AI_Challenges/AI03-Multi_Agent/Multi_Agent_App/requirements.txt
azure-search-documents>=11.0.0
azure-identity>=1.10.0
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
aiohttp>=3.8.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.20.0